Revert "(Libretro optimization) CPS3 -  Converted RamScreen from UINT32 * to UINT16 * - ~10fps speedup"
This reverts commit 281bd5ab06f933b29d5de2b69bc3654a586dde67.

+++ src-0.2.97.27/src/burn/drv/cps3/cps3run.cpp
@@ -48,7 +48,7 @@ static UINT8 *RamC000_D;
 static UINT16 *EEPROM;
 
 UINT16 *Cps3CurPal;
-static UINT16 *RamScreen;
+static UINT32 *RamScreen;
 
 UINT8 cps3_reset = 0;
 UINT8 cps3_palette_change = 0;
@@ -511,7 +511,7 @@ static INT32 MemIndex()
 	RamEnd		= Next;
 	
 	Cps3CurPal		= (UINT16 *) Next; Next += 0x020001 * sizeof(UINT16); // iq_132 - layer disable
-	RamScreen	= (UINT16 *) Next; Next += (512 * 2) * (224 * 2 + 32) * sizeof(UINT16);
+	RamScreen	= (UINT32 *) Next; Next += (512 * 2) * (224 * 2 + 32) * sizeof(UINT32);
 	
 	MemEnd		= Next;
 	return 0;
@@ -1414,7 +1414,7 @@ static void cps3_drawgfxzoom_0(UINT32 code, UINT32 pal, INT32 flipx, INT32 flipy
 
 static void cps3_drawgfxzoom_1(UINT32 code, UINT32 pal, INT32 flipx, INT32 flipy, INT32 x, INT32 y, INT32 drawline)
 {
-	UINT16 * dst = RamScreen;
+	UINT32 * dst = RamScreen;
 	UINT8 * src = (UINT8 *) RamCRam;
 	dst += (drawline * 1024 + x);
 
@@ -1638,7 +1638,7 @@ static void cps3_drawgfxzoom_2(UINT32 code, UINT32 pal, INT32 flipx, INT32 flipy
 			case 0:
 				for( INT32 y=sy; y<ey; y++ ) {
 					UINT8 * source = source_base + (y_index>>16) * 16;
-					UINT16 * dest = RamScreen + y * 512 * 2;
+					UINT32 * dest = RamScreen + y * 512 * 2;
 					INT32 x_index = x_index_base;
 					for(INT32 x=sx; x<ex; x++ ) {
 #if BE_GFX
@@ -1655,7 +1655,7 @@ static void cps3_drawgfxzoom_2(UINT32 code, UINT32 pal, INT32 flipx, INT32 flipy
 			case 6:
 				for( INT32 y=sy; y<ey; y++ ) {
 					UINT8 * source = source_base + (y_index>>16) * 16;
-					UINT16 * dest = RamScreen + y * 512 * 2;
+					UINT32 * dest = RamScreen + y * 512 * 2;
 					INT32 x_index = x_index_base;
 					for(INT32 x=sx; x<ex; x++ ) {
 #if BE_GFX
@@ -1672,7 +1672,7 @@ static void cps3_drawgfxzoom_2(UINT32 code, UINT32 pal, INT32 flipx, INT32 flipy
 			case 8:
 				for( INT32 y=sy; y<ey; y++ ) {
 					UINT8 * source = source_base + (y_index>>16) * 16;
-					UINT16 * dest = RamScreen + y * 512 * 2;
+					UINT32 * dest = RamScreen + y * 512 * 2;
 					INT32 x_index = x_index_base;
 					for(INT32 x=sx; x<ex; x++ ) {
 #if BE_GFX
@@ -1786,7 +1786,7 @@ static void DrvDraw()
 
 	if (nBurnLayer & 1)
 	{
-		UINT16 * pscr = RamScreen;
+		UINT32 * pscr = RamScreen;
 		INT32 clrsz = (cps3_gfx_max_x + 1) * sizeof(INT32);
 		for(INT32 yy = 0; yy<=cps3_gfx_max_y; yy++, pscr += 512*2)
 			memset(pscr, 0, clrsz);
@@ -1973,7 +1973,7 @@ static void DrvDraw()
 	
 	{
 		UINT32 srcx, srcy = 0;
-		UINT16 * srcbitmap;
+		UINT32 * srcbitmap;
 		UINT16 * dstbitmap = (UINT16 * )pBurnDraw;
 
 		for (INT32 rendery=0; rendery<224; rendery++) {
